refactor(CreateGymView): Extract component GradeList
